refactor: address Security Filters audit findings
- Extract SettingsSection/SettingsItem/SettingsDivider/SettingsControlRow/
  SettingsBlockTile/SettingsSubControlRow into SettingsSectionCard.kt.
  AllSettingsScreen now consumes the shared components instead of carrying
  its own private copies, and SecurityFiltersScreen drops the parallel set
  it used to declare.
- BlockedContentRow folds into SettingsItem(trailing = { CountBadge(...) }),
  removing the bespoke navigation row.
- Replace the Spinner/Switch tile mix with WarnReportsTile = SwitchTile +
  SettingsSubControlRow(stepper), avoiding the nullable-icon code smell.
- Move the sensitive-content SegmentedButtonRow out of a cramped trailing
  slot and into SettingsBlockTile so it has full-width room.
- HiddenWordsScreen: replace the eager Column.forEach with LazyColumn, and
  add Modifier.imePadding() to the Scaffold so the keyboard no longer
  covers the Add-word field.
- Move SelectableUserList from BlockedUsersScreen.kt to SecurityListsCommon
  and drop the nullable onToggle (both call sites pass one).
- BlockListTopBar now takes selectedCount: Int instead of Set<*>; its
  UserFeedState branches are exhaustive (no else -> Unit fallthrough).
- CountBadge reserves min width so 0↔n transitions don't shift the row.
- Stepper relies on Material 3's enabled handling instead of computing
  alpha by hand and tinting LocalContentColor in three places.
- Distinct icons per tile/category (Visibility, Code, VisibilityOff, etc.)
  so the same MaterialSymbols.Tag isn't reused for unrelated rows.
